백준

+++ 백준/15552 빠른 A+B.py
@@ -3,6 +3,6 @@ import sys
 N = int(input())
 for _ in range(N):
 
-    a, b = map(int, sys.stdin.readline().split())
+    a, b = map(float, sys.stdin.readline().split())
 
     print(a+b)
